Add reason message to the task object.
Task like update_router are pushed to the queue by several methods
of the AkandaL3Manager. If a task is failing it's really hard to
give it context to figure out what is going on. Adding a reason
string to the task should make it a little clearer why is running
and help the debugging phase.

Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>

Change-Id: I63938bd32c9c086ef12c5f905023dd49e50f8f81