Dispatch on tenant id and listen for RPC
Update the listening code to catch RPC casts sent to the layer
three agent so we can get router delete and update instructions
correctly.

Change the dispatching algorithm to use the tenant id instead
of the router id. Not all messages have a router id, and it is
expensive to look it up. We will add a new layer in the worker
process to manage all of the routers for a given tenant, and
that class will handle the lookups. This will also have the
benefit of ensuring that a single tenant does not tie up all of
the router processes by sending instructions to several routers
at one time.

Change-Id: I1cc843beb1a1792b21065e4259dccaaf3a431bf5