Add notification publishing code
Create a Publisher class to send messages using a separate
thread so communication with the AMQP broker does not block
the rest of the processing.

Pass a configured publisher to each worker and let the worker
manage starting and stopping it. This avoids having to pass
all of the parameters for the publisher to the worker separately.

Make the publisher optional for the worker, to simplify writing
worker tests.

Change-Id: I95d7f05dd67ebd84f20382a98807ab483971bc5c