Build notification message format properly
Let the user of the notifications.Publisher pass us a
simple dictionary, and build a real notification message
from that using the logic replicated from openstack.common.rpc.
We don't just use the library directly because it imports
eventlet all over the place, and we are trying to avoid that.

Change-Id: I342960182881a5e42116306f6997e6ce4fd7efdb