RUG should wait the retry_delay time checking if a router is alive
DHC-1656

The is_alive method catch all the exceptions returning a boolean
while the update_state method expects an exception to delay the
new request. This change fix the issue removing the try/except
block and delaying the new request based on the boolean returned
from the is_alive method.

Change-Id: I454123dd42f4b875c02afc3c714b520f8830d159
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>