Add option for ignoring some routers by id
Use a touched file in a configurable directory to indicate that a specific
router should be ignored.

If the configuration option is blank, the rug will not try to look at
the filesystem and not ignore any routers (this is how we will want to
run in production most of the time, for efficiency).

If the configuration file refers to a valid directory, its contents are
listed and used as router ids to indicate which routers should be
"ignored" and not updated based on events. No events are passed to the
router at all.

DHC-1568