Pre-emptively clean up router ports before rebooting.
In the event that an akanda instance isn't deleted
cleanly (which we've seen in certain circumstances, like
hypervisor failures), be proactive and attempt to clean up the
router ports manually.  This helps avoid a situation where the
rug repeatedly attempts to plug stale router ports into the newly
created akanda instance (and fails).