Let the VmManager deal with deleted routers
Don't skip any processing in the state machine when we think a
router has been deleted, because the instance associated with it
might not have been deleted.

When deleting, always go to StopVM, which will check the VM status
and choose the appropriate next step.

Let the VmManager deal with a missing router by adding a GONE state. In
the state machine, when we see GONE during an Alive check, delete the
router.

Change-Id: I1324f39c173eb0fd29c40a8199adfc741d941f90