Add rug-ctl router rebuild command
Add a command to forcibly rebuild a router. It works
by destroying the instance and building a new one, which is
brute-force, but easy to understand.

I also changed the worker's lookup for tenant router
managers so the tenant id is normalized cleanly to
avoid confusion since some events use '-' and some
do not (yay, consistent!).

Change-Id: I79a622b0c6cc23041da55ec2f54f79ccf463379e