Add a ClearError state
Add a ClearError state to remove the error flag from the VM before doing
anything else with it. Add the new state transition to CalcAction. We
are still protected from boot loops triggered by poll operations because
those are never processed for a router in an error state, but any update
operations or direct commands will be.

Change-Id: I6bf744a33df8f4055a744b5fe2c7ef3d3043c247