Fix state of router when we notice it is missing
We used to rely on receiving an event to know that a router had been
deleted. We're going to be filtering that event out (see subsequent
changesets), so we need to properly detect that a router VM is gone and
mark ourselves as down so the automaton can reboot us.

Change-Id: I287a8dcc208bc709687cf63a0f17ace25265d178