Ignore POLL requests when in an error state
There is a race condition between setting a router into an error state
and receiving a poll request from the health monitor. If we do see a
poll request, ignore it and no-op the transition back to CalcAction
instead of resetting an error state that may have been set after the
poll action went into the queue.

Change-Id: Ie82d937812b79357bdcda13dd11aa3a3d732a0dc