Add NoopPublisher
The code that enables publishing notifications is spread across several
code paths. Instead of checking for the toggle flag at every point or in
the Publisher class itself we replace the Publisher class with a
NoopPublisher class which simply stubs out the methods and does nothing.

This allows us to configure and declare the behaviour in main.py without
spreading more conditionals all over the place.