Fix a race condition that affects router VM boots.
In Icehouse, the `update_router` L3 API command changed so that the
process of setting (and potentially creating the gateway port) is no
longer synchronous and part of the database update.  Because of this, we
can no longer rely on calls of `update_router` to include port
information (because it may not have been created yet).  Instead, we need to poll a few times with `show_router` until the gateway port has been
created:

https://github.com/openstack/neutron/commit/f23f2ecee68ba4abd12139bbb91b77ba9410f581