Assign management port to a management security group
This adds creation of a management security group on the service tenant
that ensures IPv4 + IPv6 API traffic can reach the appliance. It also
enables SSH and ICMP by default, for debugging.

Additionally, it adds a new configuration setting
'management_extra_security_group_ids' that allows operators to specify
additional security groups to be assigned to the management port.

Closes-bug: #1451622

Change-Id: I5a47cd306f6f7f1b005c26971a042163a2f86e88