Add tools/run_functional.sh
Adds a place holder for what will be the entry point into the akanda
functional test suite.  devstack-gate will call this script via a
post_test_hook after it's spun up a devstack-backed cloud.  This is
a noop for now until the devstack-gate glue is in place and we have
something we can test.

Change-Id: Ieb25645801a32f1262ddb2967e834755f813892a