Remove appliance's hard-coded password in favor of ssh pub key
This removes the hard-coded password from the appliance VM's user-data
and replaces it with a SSH public key, which is read from a file whos
path is configured in rug.ini.

It also disables password logins for the user.  Another patch to
akanda-appliance-builder will allow developers to include a specified
debug user to allow debugging in dev environments.

Change-Id: I7db92bc7fd3743d89d73ab2a0b8da14685c30c69