Make blessed an optional dependency
blessed is not in openstack g-r and blocks us from aligning there. This makes
the dependency optional.  If it is not installed, the cli browse command emits
a warning telling the user to install it.  The browse section of the CLI is the
only place its used and doesn't affect any other functionality.

Change-Id: If446c2135936c8ea4083eb9fc712a98031b28ac4