Wait for a configured test router to become active
The router VM that is spawned during RUG startup may not be active until some
time after the test suite, causing tests to start before the thing is active.
This adds an assertion to the functional test setUp to wait for a configured
router to become active before continuing on with any tests.  The router to
wait on is now configured in test.conf.  run_functional.sh will attempt to find
the devstack created one, and error if there are multiples.  Users can specify
their own test router if running outside of the gate env.

Change-Id: I2b54ddfcce8dab446ac7418cca61320b370bc999