Repair the semi-functional notification publisher
This begins to fix the RUG's bandwidth notification code.  If configured
to use ceilometer and you somehow trigger a ReadStats state transition,
you'll find that the vm manager does not expose any way of collecting
stats.  This adds some basic functionality that we can refine to meet
standard ceilometer and neutron notification specs.

Change-Id: I17032c6f9e4a7f3a9f575a1667498c1cc79eba4c