Migrate to oslo.log
This updates logging to use olso.log. Also updates the default logging format
to be a bit more standard, and sets up devstack to configure colorized logging
in the screen session.

This will likely need to be squashed into the oslo.messaging patch but pushing
seperately for the sake of initial review.

Change-Id: If8353c96b0a6d5281aff2a226f54f5aa2233acb3