Fix ability to recover from an existing appliance VM
This fixes two issues discovered when trying to reboot a devstack
host, which is similar to a full datacenter reboot.

First, the RUG currently tries to delete any existing non-alive VMs
it finds for a router.  Its failing to do that ATM, attempting to
use an ID from the wrong object.

With that fixed, the VM manager and state machine are still incapable of
removing and starting with a fresh appliance VM.  Currently, the pass
through the state machine that detects and deletes the existing VM counts
as a negative boot attempt and the execution does not clear its cached
instance_info after its deleted the instance.  This causes another pass
through the state with a stale instance_info that again counts against
the boot count.  That causes the router to get set to ERROR and future
POLL events to get dropped, leading to no replacement router to booting.

Change-Id: I7d1a0a58886088cf279a68b5aaf4cff2a678e16a
Closes-bug: #1468045