Migrate to oslo.log and oslo.messaging
This was originally proposed as two separate patches but a module dependency
issue requires them to be squashed into one patch.

* Migrate to oslo.log

This updates logging to use olso.log. Also updates the default logging format
to be a bit more standard, and sets up devstack to configure colorized logging
in the screen session.

This will likely need to be squashed into the oslo.messaging patch but pushing
seperately for the sake of initial review.

* Rework messaging layer for oslo.messaging

Instead of setting up kombu connections directly this uses oslo.messaging.
oslo.messaging doesn't expose all the kombu/rabbit-isms that let us setup a
single connectino to sniff the notifications and RPC buses.  Instead, we need
to subscribe to each as different connectinos. The child process that listens
for notifications now spawns multiple threads, one for each consumer we setup.

This also updates the Sender class to use an oslo.messaging notifier instead of
setting up notifications for commands directly with kombu. The ceilometer
Publisher uses the same approach when sending its messages.

The old code that created Events from messages is now split between a
notifications endpoint object and another for RPC messages.

Change-Id: Ia7cc26dc20ed57bdf21571e260d006669d6f3310