devstack: set use_ipv6=True in nova.conf, workaround nova bug
Set config setting that is now required to ensure metadata can be constructed
correctly for ipv6 networked instances.

Also, pull down an in-flight patch to nova from
https://review.openstack.org/#/c/203846/ that fixes a nasty
bug in Nova that blocks akanda appliance boot.  This is temporary and will
need to be removed once that fix lands in Nova proper.

Change-Id: I5b9911f7393c1cc2f75da53504cffcb67215e77b