Adds a DB layer, use it for debug modes
In preparation for scale out RUG, this adds a database layer
(built on oslo.db) that will be used for managing debug modes.
Instead of tracking debug'd/ignored routers and tenants in-memory
or on disk as files, this uses a database. This means that putting
things into debug mode via rug-ctl are now persistent, and the file-based
approach is no longer available.  A sqlite database (the default) can be
used for single node installs, or the RUG can be pointed at mysql/pg to
handle this in larger environments.

This also adds a global debug mode that can be used to ignore all events
during maintanence periods.

A new optional 'reason' argument has been added to the debug modes, allowing
operators to add a note when entering a tenant/router/cluster into debug
mode.

Change-Id: I3f5129e11b11cf5aaed8889da3b204104e5ad203
Closes-bug: #1470619
Partially implements: blueprint rug-scaling