Move caching of resolved router_ids from state machine to worker
If an event comes in that does not have a associated router_id, we attempt
to call out to neutron to resolve it by tenant_id and cache it locally on
the TenantRouterManager instance. This happens during get_state_machines()
and, if a router could not be found, the event msg effectively gets dropped.

This is a bit of a waste of Neutron calls if the event happens to be one
that would be ignored later on by the worker (for global or tenant debug,
for example)

This moves that caching out to the worker using a dumb in-memory dict based
cache (we could back this by memcache and improve it if it ends up being
beneficial)  Instead of having the TRM decide to drop messages with no
associated router, we move that responsibility up to the worker where its
already doing similar things around debug modes.  Ensuring the worker has
visibility of the router_id at all times will later let it use this as a
hash key when RUGs are managed using a distributed hash ring.

Partially implements: blueprint rug-scaling

Change-Id: Iaf2ded91c2c777855f874ac01a54e96c9045e29c