Track last_boot from nova server not instance_info creation
We currently only ever get an InstanceInfo's last_boot timestamp by
creating it on instantiation of an InstanceInfo object. This does
not scale to multiple Rugs and can cause quirks on cold startup of
multiple Rugs. This changes it to instead derive the last_boot
timestamp from the 'created_at' field of the Nova server.

It also removes an unused 'instance_up' attribute and adds a helper
class method for creating the InstanceInfos from novaclient objects.

Change-Id: I882412538e74cd75f31d9eb124f156cc16bfd46f