Fix pylint warnings
Fixes some of the more interesting pylint warning
messages. Most changes were for things that were
potentially dangerous, though there are a couple
(i.e. not breaking out of a for loop with an else
condition) that could cause unexpected behavior.

This does not address all pylint warnings. There
are still many that are not legitimate concerns,
though they could be reviewed for other possible
problems not caught by this patch.

Change-Id: I5decd705ef7ace482fb2632172ae43eec04f8911