Enrich functional test suite
This beefs up the functional test suite to do some tests on a
tenant router.  The test can now creates one-off tenants to be used
in the tests.  The new tests act entirely on behalf of the test
router and attempt on only use admin-level things when required.

Partially-Implements: blueprint ci-updates-mitaka

Depends-On: I41c4a49b81d62b0caddc9270b3f8e4a9315d1172
Change-Id: I26fa70f877522f09075dc87583f2359cc0dcaf41