Properly detect backing instance is gone
The instance_manager's current caching of instance_info only happens
once and is never refreshed.  This causes update_state() to not notice
a backing instance is gone until it fails a API liveliness check.

This fixes it by updating the instance_info on every run of ensure_cache(),
in the same way it does for the router_obj.  This will properly detect
a backing instance is gone and rebuild the router without the slow
failing liveliness check.

Fixing this introduces a race where we try to check the API server of
a booted instance that does not yet have ports associated.  This also
fixes this by considering instances with no management port to still
be in the booting state.

Change-Id: I847b8a15239c7a195a0593ca50344ea7d5fe35c2
Closes-bug: #1492056