devstack: Update for keystone v3 auth
This updates the devstack plugin to be able to use CLI client auth params
that are appropriate for the current keystone auth version.  This'll
let our plugin continue to function with v3 auth, which is now the default
in devstack.

This does not update the rug itself to be v3 compat, we'll need to address
that in another patch.

Change-Id: Ie8560c91c4666ed332258d9a917cb03b05ea9fae
Closes-bug: #1475781