Split out instance provisioning to on_demand and pez (WIP)
This splits up the ways in which the Rug may request an instance
to be used for an appliance.  We preserve existing behavior of
booting a new instance on-demand as the on_demand instance provider
and introduce a new pez instance provider that requests already-booted
appliance VMs from the Pez service.

Partially-implements: blueprint nodepool-support

Change-Id: I00b3f5e3f8e6b65b860acce7ab1772635d9cdf6d