Start of the LBAAS driver
This is the start of the LBAAS driver.  If you point [lbaas] config
at the existing akanda appliance image, it should spin up a new instance
on and manage the neutron resource's state according to the akanda resource.

This LB driver expects to be able to push its config to the akanda-appliance
server running in the image, and that server is responsible for translating
what the RUG gives it to something suitable for the LB service installed in
that image.  There will be a corresponding change to akanda-appliance for this.

The port creation currently creates a instance port on the same network as the
LB VIP.  I'm not sure what else we need to do in terms of port plugging.

List of other things that have been done:

    - Split akanda.rug.api.configuration into akanda-rug.api.config.{common,
      router} in prep for building LBAAS configs as well.
    - Build configuration for both routers and LBs using their own respective
      config libraries, but push to the same appliance via the /v1/system/

TODO:
    - Figure out correct port plugging
    - Subscribe to other LBAAS events and translate those into config updates

This requires a corresponding change to akanda-neutron to enable
the akloadbalancerstatus api extension.

Change-Id: Ic19a883f56fb6d65a83b1f4d93b581f9e242d97f