Flag initial cluster rebalances as bootstrap events
We use ther rebalance event for both reacting to remote cluster changes
and initially seeding the workers hash ring with the peer nodes when it
comes online for the first time. Its required that we do some altering of
the local state in reaction to remote cluster events, but in the case of
the initial bootstrapping event, this isn't necessary as there isn't any
local state. This avoids messing with the TRMs and local state machines
during bootstrapping, as it ends up creating dual-TRMs later that cause
multiple workers to end up acting on a single resource.

Change-Id: If00fbadfef4b327e72b3ebd1aa83b6963df561a4