Introduces the Pez instance provider
This abstracts away how/where akanda.rug.api.nova.boot_instance()
creates a new resource.  It preserves existing behavior of booting
one-off instances for resources in the OnDemandInstanceProvider.

It introduces a PezInstanceProvider which relies on a new service
akanda-pez-service.  This service is responsible for keeping pools
of hot-standby nodes to be ready for use as appliance VMs.  The size
of said pools are configurable and it maintains separate pools for
each enabled driver type.

When configured with the the PezInstanceProvider, the Rug will request
instances via RPC from Pez instead of booting directly from Nova.

To enable, you must run the akanda-pez-service and then set
'instance_provider=pez' in rug.ini.  The default, which preserves existing
behavior, is 'instance_provider=on_demand'

Implements: blueprint nodepool-support

Change-Id: I7680c747d8c351c63d675f5fb377eee5edbf7b31