Dynamically allocate service port addresses
We currently hard-code the address for management and external
ports to the first address on the subnet, which breaks clustering
astara-orchestrators when two nodes attempt to create neutron ports
with the same addresses.  This updates usage to instead rely on
a Neutron-assigned address to bring up locally.

Note this is a partial fix: we'll need corresponding changes that
allow us to push in this address to appliances for metadata access,
which is hard-coded there as well to the same address.

Change-Id: I88fa97bae84ca245afa5ad0da4ac3c0bc1c441ff
Partial-bug: #1524068