Push orchestrator config into the appliance
This pushes a couple of flags into the appliance that are specific to the
individual orchestrator instance managing that appliance. Initially, we use
it to tell the appliance where the metadata proxy is listening.  Previously,
this was hard-coded to a known address on the network.  With multiple
orchestrators in a clustered env, this will allow each to run their own
metadata proxy and have only their managed appliances querying that.

Another patch will follow that will ensure this is up to date when rebalances
occur and orchestrators take over new appliances.

Change-Id: Ib502507b29f17146da81f61f34957cd96a1548f4
Partial-bug: #1524068