Cleanup unmanaged state machines via its own code path
We're reusing the resource deletion code to cleanup unmanaged
resources.  This assumes the resource has been deleted from Neutron
and sets a flag accordingly. This preventing an orchestrator from regaining
control of a resource if it has failed over to another node, and then back
to its original orchestrator again.  Instead, introduce a new code path
for unmanaging a resource that does not flip its deleted flag, allowing the
state machine to be recreated the next its needed.

Change-Id: I76bf702c33ac6ff831270e7185a6aa3fc4c464ca
Closes-bug: #1527396