Cleanup SM management during rebalance events.
This cleans up the worker's handling of rebalance events a bit
and ensures we dont drop state machines in a way that prevents
them from later being recreated.  It also avoids a bug where, upon
failing over resources to a new orchestartor, we create a state
machine per worker, instead of dispatching them to one single worker.

To do this, the scheduler is passed into workers as well as the
process name, allowing them to more intelligently figure out what
they need to manage after a cluster event.

Finally, this ensures a config update is issued to appliances after
they have moved to a new orchestrator after a cluster event.

Change-Id: I76bf702c33ac6ff831270e7185a6aa3fc4c464ca
Partial-bug: #1524068
Closes-bug: #1527396