Adds a new rebalance takeover state
This adds a new state to the graph that is triggered for resources
that have been migrated to be managed by the local orchestrator after
a rebalance event.  It adds a new driver method for drivers to express
what needs to happen to update their appliance after such an event.

The driving case for this is to be able to update configuration for
astara-appliance routers after a rebalance, in order to update its
metadata proxy configuration.  Some of the retry logic around config
update has been moved out of the instance manager and into the astara
client library.

Partial-bug: #1524068

Change-Id: I25739d58791e36f16541219f1dfa238ae229eeae