Cleanup deleted resource from the tenant resource cache
When a resource is deleted, its not currently removed from the tenant
resource cache.  This causes a cache hit if the tenant attempts to re-create
the same type of resource, but the resource is then later ignored because
it has been deleted.  This adds a callback used by the TRM to remove it
from the resource cache when its state machine is deleted.

Change-Id: I5dcbeda7de240a693fc7a4944dd34a37b10d174b
Closes-bug: #1531597