Stop repeatedly logging boot time
We uncoditionally log the boot time of a resource every time update_state is
called. This causes an increasing boot time to be logged every time the
orchestrator is polled or a health check fired.  This should only be logged
once, so use a flag on the object to suppress future logging.

Change-Id: I18dc97482c9e09253bf7e692a3ed0857ac8308d6