Drop orchestrator.ini.sample from git
This drops tracking of the generated sample from git, since the output
of generation is non-deterministic, we end up with big changes to this
file anytime someone proposes a change that requires a new sample.

Instead, lets track orchestrator.ini and update it at milestone/release
times.

Also, have the devstack plugin generate and use a sample, that way we
get some gate testing on proposed config changes.

Change-Id: Iaf5127733765e973dbf0e812f267a3304c575fd4
Closes-bug: #1535889