Correct state transition from stop when instance is gone
Correctly mark the state and return it from stop() when the instance
has already been deleted.  This makes the state machine correctly
recover from situations where nova instances disappear when
a config update is in-flight.  It also ensures delete_ports()
is called on the driver in all cases, to ensure port cleanup.

Change-Id: I41c4a49b81d62b0caddc9270b3f8e4a9315d1172
Closes-bug: #1535961