Remove external network hooks and auto-added resources
Previously the orchestrator could be configured to create access to the
external network.  This combined with auto external gateway additions
would cause problems with automation tools and did not work reliably.
This change removes this functionality since it was often disabled in
production deployments. This change aslo slightly reduces devstack
runtime by removing the Neutron restart.

Change-Id: I556f1fc2729f1d62a60de24b6d5e9ed473749f9a