Remove redundant config push in CheckBoot()
The CheckBoot state should really only be concerned with testing that the
instance has successfully booted and its API is up and ready for
ConfigureInstance's config push.  This is handled entirely by update_state()
and pushing the config here as part of the boot test is redundant, which can
be problematic in environments where a config update is an expensive task.

Change-Id: I5de24c627e468ba12fbccd3a3be713db380b645a