Handle a race between pre-populate and hash ring bootstrapping
On startup, if the worker receives messages for pre-populated resources
prior to processing the initial cluster rebalance event, the messages
will be dropped. This fixes the race by tracking when the hash ring
has been initialized.  Any events it receives prior to finishing init
will be batched up and processed as part of the initial bootstrapping
procedure.

Change-Id: I3caf95f57380076ab48e4270e1cd575906fba386
Closes-bug: #1554248