Include network config in loadbalancer config push
In order for LBs to know about their default gateways, we need to include
network_config in the 'loadbalancer' bucket of the config push.  This requires
we tailor it to only include the single fixed ip on the subnet it was created,
instead of fixed ips for all subnets on the network.

Partial-bug: #1568924

Change-Id: I49d5932890ac8032c2743d79bd4b06c4655c794f