Make Astara Newton compatible
This change makes Astara compatible with Newton. This change adds support
for Neutron agent reporting and service providers to work with new L3
drivers.  Included in this change is a temporary filed called newton_fix.py.
This file will be removed in follow-up change after changes are migrated to
astara-neutron.

Change-Id: I5843e84e36af2e46de5b8420ca5749033c26ee69